검색 메인 화면에서 검색어를 입력 후 버튼을 누르면 결과 화면으로 이동하도록 수정
## 수정 사항
- 검색 메인 화면에서 검색어를 입력 후 버튼을 누르면 결과 화면으로 이동하도록 수정
- 검색 메인 화면에 필요한 검색창 구성 View를 SearchBoxView로 분리
- contentview에서 화면을 구성하던 코드를 SearchMainView로 이동

## 추후 작업해야 하는 부분
- 결과 화면에서 탭바로 뉴스와 블로그를 보여주도록 설정
- 결과 화면의 List가 가운데로 이동하는 현상 수정
- 검색 화면에서 키보드 엔터로 이동할 수 있도록 수정
